# assets/schemas/storage-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Obelisk storage API",
    "version": "1",
    "description": "Authenticated operation-level storage protocol. Positional JSON arguments and typed Result responses. Notification polls last at most 30 seconds; responses and finished take a maximum wait in milliseconds as their last argument."
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/blobs": {
      "post": {
        "operationId": "Cas_write_blob",
        "tags": [
          "Cas"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Typed database result; Err may follow a committed write and must not be replayed automatically.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_of_string_or_WireError"
                }
              }
            }
          },
          "400": {
            "description": "Malformed arguments"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Unknown operation or protocol version"
          },
          "413": {
            "description": "Request body exceeds 128 MiB"
          },
          "503": {
            "description": "Request or long-poll capacity exhausted"
          },
          "504": {
            "description": "Request exceeded the 60-second deadline; a write may have committed"
          }
        }
      }
    },
    "/v1/blobs/{digest}": {
      "parameters": [
        {
          "name": "digest",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "Cas_read_blob",
        "tags": [
          "Cas"
        ],
        "responses": {
          "200": {
            "description": "Blob bytes",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid digest"
          },
          "401": {
            "description": "Invalid bearer token"
          },
          "404": {
            "description": "Blob absent"
          },
          "500": {
            "description": "Database read failed"
          },
          "503": {
            "description": "Storage unavailable"
          },
          "504": {
            "description": "Request deadline exceeded"
          }
        }
      },
      "head": {
        "operationId": "Cas_contains_blob",
        "tags": [
          "Cas"
        ],
        "responses": {
          "200": {
            "description": "Blob exists"
          },
          "400": {
            "description": "Invalid digest"
          },
          "401": {
            "description": "Invalid bearer token"
          },
          "404": {
            "description": "Blob absent"
          },
          "500": {
            "description": "Database read failed"
          },
          "503": {
            "description": "Storage unavailable"
          },
          "504": {
            "description": "Request deadline exceeded"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "WireError": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "Closed",
              "NotFound",
              "Conflict",
              "AlreadyFinished",
              "StubConflict"
            ]
          },
          {
            "type": "object",
            "properties": {
              "Generic": {
                "type": "string"
              }
            },
            "required": [
              "Generic"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "ValidationFailed": {
                "type": "string"
              }
            },
            "required": [
              "ValidationFailed"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "IllegalState": {
                "type": "string"
              }
            },
            "required": [
              "IllegalState"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "UnlockedCannotBeAppended": {
                "type": "string"
              }
            },
            "required": [
              "UnlockedCannotBeAppended"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "VersionConflict": {
                "type": "object",
                "properties": {
                  "expected": {
                    "type": "integer",
                    "format": "uint32",
                    "minimum": 0
                  },
                  "requested": {
                    "type": "integer",
                    "format": "uint32",
                    "minimum": 0
                  }
                },
                "required": [
                  "expected",
                  "requested"
                ]
              }
            },
            "required": [
              "VersionConflict"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "SubscriptionEnded": {
                "$ref": "#/components/schemas/ResponseSubscriptionEnd"
              }
            },
            "required": [
              "SubscriptionEnded"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "Timeout": {
                "$ref": "#/components/schemas/TimeoutOutcome"
              }
            },
            "required": [
              "Timeout"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "Cas": {
                "type": "string"
              }
            },
            "required": [
              "Cas"
            ],
            "additionalProperties": false
          }
        ]
      },
      "ResponseSubscriptionEnd": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "PollIntervalElapsed",
              "ExecutorClosing",
              "PauseOrCancel"
            ]
          },
          {
            "type": "string",
            "const": "LockDeadlineReached",
            "description": "Execution run should return without writing any more events, execution remains blocked waiting for response"
          }
        ]
      },
      "TimeoutOutcome": {
        "type": "string",
        "enum": [
          "Timeout",
          "Cancel"
        ]
      },
...
      "Result_of_string_or_WireError": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "Ok": {
                "type": "string"
              }
            },
            "required": [
              "Ok"
            ]
          },
          {
            "type": "object",
            "properties": {
              "Err": {
                "$ref": "#/components/schemas/WireError"
              }
            },
            "required": [
              "Err"
            ]
          }
        ]
      }
    }
  }
}